backend/requirements.txt
fastapi
google-generativeai
graphviz
matplotlib
numpy
openai
pandas
pm4py
pydantic
python-multipart
seaborn
uvicorn
httpx
PyJWT
google-adk
google-cloud-storage
motor
